requirements.txt
pyyaml
pydantic
typer>=0.9.0
numpy
openai
tiktoken
timeout-decorator